requirements.txt
wget
wheel
psutil
requests
tgcrypto
ffmpeg-python
python-dotenv
youtube_search
youtube_search_python
pyrogram
pytgcalls==3.0.0.dev24